# swagger.yaml
openapi: 3.0.0
info:
  title: Revo
  contact: {}
  version: '1.0'
servers:
  - url: https://api-dev.payloow.com/api
  - url: http://localhost:4000/api
  - url: http://localhost:8000/api
    variables: {}
paths:
  /v1/register:
    post:
      tags:
        - User
      summary: register user
      operationId: registeruser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/registeruserrequest'
                - example:
                    firstName: Wisdom
                    lastName: Dakoh
                    password: wisdom
                    phone: '2347052316803'
                    email: wisdomdakoh@gmail.com
                    country: Nigeria
                    state: Lagos
                    role: user
                    address: 123 Example Street, Victoria Island
            example:
              firstName: Wisdom
              lastName: Dakoh
              password: wisdom
              phone: '2347052316803'
              email: wisdomdakoh@gmail.com
              country: Nigeria
              state: Lagos
              role: user
              address: 123 Example Street, Victoria Island
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...
  /v1/login:
    post:
      tags:
        - user Auth
      summary: login user
      operationId: loginuser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/loginuserrequest'
                - example:
                    phone: '2347052316803'
                    password: wisdom
            example:
              phone: '2347052316803'
              password: wisdom
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...
  /v1/forget-password:
    post:
      tags:
        - user Auth
      summary: forget password
      operationId: forgetpassword
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/forgetpasswordrequest'
                - example:
                    email: marvelloussolomon49@gmail.com
                    deviceType: mobile
            example:
              email: marvelloussolomon49@gmail.com
              deviceType: mobile
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /v1/reset-password:
    post:
      tags:
        - user Auth
      summary: reset password
      operationId: resetpassword
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/resetpasswordrequest'
                - example:
                    email: marvelloussolomon49@gmail.com
                    token: a3091dda44767000836f2d8cd768f0e0
                    password: marvel12
                    code: 123456
            example:
              email: marvelloussolomon49@gmail.com
              token: a3091dda44767000836f2d8cd768f0e0
              password: marvel12
              code: 123456
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...
  /v1/user/dva:
    get:
      tags:
        - User
      summary: Get user DVA information
      operationId: getUserDva
      parameters: []
      responses:
        '200':
          description: Successfully retrieved user DVA information
          headers: {}
      deprecated: false
      security:
        - bearer: []
  /v1/set-transaction-pin:
    post:
      tags:
        - user Auth
      summary: Set Transaction pin
      operationId: SetTransactionpin
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/SetTransactionpinRequest'
                - example:
                    pin: '1234'
            example:
              pin: '1234'
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /v1/update-transaction-pin:
    patch:
      tags:
        - Settings
      summary: Update transaction PIN
      operationId: updateTransactionPin
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/updateTransactionPinRequest'
                - example:
                    currentPin: '1234'
                    newPin: '5678'
            example:
              currentPin: '1234'
              newPin: '5678'
        required: true
      responses:
        '200':
          description: Successfully updated transaction PIN
          headers: {}
      deprecated: false
      security:
        - bearer: []
...
  /v1/buy-data:
    post:
      tags:
        - Bill Payment - Gongozo
      summary: Buy Data
      operationId: BuyData
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/BuyDataRequest'
                - example:
                    network: 1
                    mobile_number: '09095263835'
                    plan: 364
            example:
              network: 1
              mobile_number: '09095263835'
              plan: 364
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...
  /v1/airtime-to-cash:
    post:
      tags:
        - Artime to cash with bingPay
      summary: airtime to cash
      operationId: airtimetocash
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/airtimetocashrequest'
                - example:
                    amount: 5000
                    network: 1
                    phone: '07061785113'
            example:
              amount: 5000
              network: 1
              phone: '07061785113'
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
  /v1/network-fee:
    post:
      tags:
        - Artime to cash with bingPay
      summary: network fee
      operationId: networkfee
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/networkfeerequest'
                - example:
                    amount: 5000
                    network: 4
            example:
              amount: 5000
              network: 4
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...
components:
  schemas:
...
    registeruserrequest:
      title: registeruserrequest
      required:
        - firstName
        - lastName
        - password
        - phone
        - email
        - country
        - state
        - address
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        password:
          type: string
        phone:
          type: string
        email:
          type: string
        country:
          type: string
        state:
          type: string
        address:
          type: string
      example:
        firstName: john
        lastName: Doe
        password: wisdom
        phone: '2348001234562'
        email: john.do@exampl.com
        country: Nigeria
        state: Lagos
        address: 123 Example Street, Victoria Island
...
    loginuserrequest:
      title: loginuserrequest
      required:
        - phone
        - password
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
      example:
        phone: '2347052316803'
        password: wisdom
    forgetpasswordrequest:
      title: forgetpasswordrequest
      required:
        - email
      type: object
      properties:
        email:
          type: string
      example:
        email: marvelloussolomon49@gmail.com
    resetpasswordrequest:
      title: resetpasswordrequest
      required:
        - email
        - token
        - password
      type: object
      properties:
        email:
          type: string
        token:
          type: string
        password:
          type: string
      example:
        email: marvelloussolomon49@gmail.com
        token: a3091dda44767000836f2d8cd768f0e0
        password: marvel12
...
    SetTransactionpinRequest:
      title: SetTransactionpinRequest
      required:
        - pin
      type: object
      properties:
        pin:
          type: string
      example:
        pin: '1234'
...
    BuyDataRequest:
      title: BuyDataRequest
      required:
        - network
        - mobile_number
        - plan
      type: object
      properties:
        network:
          type: integer
          format: int32
        mobile_number:
          type: string
        plan:
          type: integer
          format: int32
      example:
        network: 1
        mobile_number: '09095263835'
        plan: 364
...
    airtimetocashrequest:
      title: airtimetocashrequest
      required:
        - amount
        - network
        - phone
      type: object
      properties:
        amount:
          type: integer
          format: int32
        network:
          type: integer
          format: int32
        phone:
          type: string
      example:
        amount: 5000
        network: 1
        phone: '07061785113'
    networkfeerequest:
      title: networkfeerequest
      required:
        - amount
        - network
      type: object
      properties:
        amount:
          type: integer
          format: int32
        network:
          type: integer
          format: int32
      example:
        amount: 5000
        network: 4
...
    updateTransactionPinRequest:
      title: updateTransactionPinRequest
      required:
        - currentPin
        - newPin
      type: object
      properties:
        currentPin:
          type: string
        newPin:
          type: string
      example:
        currentPin: '1234'
        newPin: '5678'
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
    apiKey:
      type: apiKey
      name: Authorization
      in: header
security:
  - bearer: []
...